divide main.dart into several files/class/modules
