+++ lib/navigation_bar/navbar_item.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class NavBarItem extends StatelessWidget {
+  final String title;
+  const NavBarItem(this.title);
+
+  @override
+  Widget build(BuildContext context) {
+    return Text(
+      title,
+      style: TextStyle(fontSize: 18),
+    );
+  }
+}
