+++ lib/navigation_bar/navbar_logo.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class NavBarLogo extends StatelessWidget {
+  const NavBarLogo({Key key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      height: 80,
+      width: 150,
+      child: Image.asset('assets/logo.png'),
+    );
+  }
+}
