+++ pubspec.yaml
@@ -24,6 +24,7 @@ dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.2
+  responsive_builder: ^0.4.1
   url_launcher: ^6.0.7
 
 dev_dependencies:
